+++ Student-Management-System/src/main/java/com/easyWay/Student_Management_System/Entity/FacultySalaryEntity.java
@@ -10,13 +10,13 @@ import java.util.UUID;
 @Entity
 @Data
 public class FacultySalaryEntity extends BaseEntity{
-    private int facultySalary;
-    private int facultyTax;
-    private int facultyTransport;
+    private float facultySalary;
+    private float facultyTax;
+    private float facultyTransport;
 
     @Column(columnDefinition = "Text")
     private String facultyDeduction;
-    private int total;
+    private float total;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "faculty_info_id", nullable = false)
